Updated tests to use native ldap methods when not testing wrapper methods

+++ tests/ConnectionTest.php
@@ -23,6 +23,26 @@ class ConnectionTest extends TestCase
         ];
         $this->dn = sprintf('cn=%s,%s', $this->entry['cn'], $base);
         $this->filter = sprintf('(&(objectClass=virtualMachine)(cn=%s))', $this->entry['cn']);
+
+        $host = $config['host'];
+        $port = isset($config['port']) ? $config['port'] : 389;
+        $bind_dn = isset($config['bind_dn']) ? $config['bind_dn'] : null;
+        $bind_pw = isset($config['bind_pw']) ? $config['bind_pw'] : null;
+
+        $this->canonical_conn = ldap_connect($host, $port);
+        if ($this->canonical_conn === false) {
+          throw new \Exception('Can\'t connect to LDAP server');
+        }
+
+        $result = ldap_set_option($this->canonical_conn, LDAP_OPT_PROTOCOL_VERSION, 3);
+        if (!$result) {
+            throw new \Exception('Can\'t request LDAPv3');
+        }
+
+        $result = ldap_bind($this->canonical_conn, $bind_dn, $bind_pw);
+        if (!$result) {
+            throw new \Exception('Can\'t bind to LDAP');
+        }
     }
 
     /**
@@ -79,21 +99,21 @@ class ConnectionTest extends TestCase
         }
 
         $this->assertEquals($limit, $i);
+
+        $result = ldap_delete($this->canonical_conn, $this->dn);
+        if (!$result) {
+            throw new \Exception('Can\'t delete');
+        }
     }
 
     public function testDelete()
     {
-        $limit = 1;
-        $search_results = $this->conn->search(Connection::BASE, $this->dn, $this->filter, [], $limit);
-        $i = 0;
-
-        foreach ($search_results as $dn => $attrs) {
-            $this->assertNotEquals('', $dn);
-            $this->assertArrayHasKey('count', $attrs);
-            $i++;
+        $result = ldap_add($this->canonical_conn, $this->dn, $this->entry);
+        if (!$result) {
+            throw new \Exception('Can\'t add');
         }
 
-        $this->assertEquals($limit, $i);
+        $limit = 0;
 
         $this->conn->delete($this->dn);
 
@@ -105,7 +125,7 @@ class ConnectionTest extends TestCase
                 $i++;
             }
 
-            $this->assertEquals(0, $i);
+            $this->assertEquals($limit, $i);
         } catch (\Exception $e) {
             $expected_code = 0x20;
             $error_code = $e->getCode();
@@ -119,19 +139,10 @@ class ConnectionTest extends TestCase
         $test_dn_rename_full = sprintf('%s,%s', $test_dn_rename, $this->base);
         $test_filter_rename = sprintf('(&(objectClass=virtualMachine)(%s))', $test_dn_rename);
 
-        $this->conn->add($this->dn, $this->entry);
-
-        $limit = 1;
-        $search_results = $this->conn->search(Connection::BASE, $this->dn, $this->filter, [], $limit);
-        $i = 0;
-
-        foreach ($search_results as $dn => $attrs) {
-            $this->assertNotEquals('', $dn);
-            $this->assertArrayHasKey('count', $attrs);
-            $i++;
+        $result = ldap_add($this->canonical_conn, $this->dn, $this->entry);
+        if (!$result) {
+            throw new \Exception('Can\'t add');
         }
-
-        $this->assertEquals($limit, $i);
 
         $this->conn->rename($this->dn, $test_dn_rename, null, true);
 
@@ -163,7 +174,10 @@ class ConnectionTest extends TestCase
             $this->assertEquals($expected_code, $error_code);
         }
 
-        $this->conn->delete($test_dn_rename_full);
+        $result = ldap_delete($this->canonical_conn, $test_dn_rename_full);
+        if (!$result) {
+            throw new \Exception('Can\'t delete');
+        }
     }
 
     public function scopeProvider()
@@ -192,6 +206,7 @@ class ConnectionTest extends TestCase
     public function tearDown()
     {
         unset($this->conn);
+        ldap_unbind($this->canonical_conn);
     }
 }
